Use print instead of puts for logging
I observed some instances of a single line receiving two sha/event
pairs, which broke JSON parsing. One possible explanation for this is
that internally puts uses two system calls - one to write the argument,
and one to write the newline - and it seems possible that if multiple
processses are writing to the same log, you could see issues like this.
Writing both the message and the newline in the same call should prevent
this issue.